log in page is fully styled
